requirements.txt
django==5.2.6
djangorestframework==3.16.1
psycopg2-binary==2.9.10
gunicorn==21.2.0